[build] Build-logic refactoring
