+++ build-logic/convention/src/main/kotlin/com/moya/funch/plugins/AndroidFeaturePlugin.kt
@@ -1,12 +1,25 @@
 package com.moya.funch.plugins
 
+import com.moya.funch.plugins.utils.configureAndroidCommonPlugin
+import com.moya.funch.plugins.utils.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.kotlin.dsl.apply
+import org.gradle.kotlin.dsl.dependencies
+import org.gradle.kotlin.dsl.kotlin
 
 class AndroidFeaturePlugin : Plugin<Project> {
-    override fun apply(target: Project) =
+    override fun apply(target: Project) {
         with(target) {
-            plugins.apply("com.android.library")
-            configureAndroidCommonPlugin()
+            apply<AndroidLibraryPlugin>()
+
+            dependencies {
+                add("implementation", libs.findLibrary("core.ktx").get())
+                add("implementation", libs.findLibrary("appcompat").get())
+                add("implementation", libs.findBundle("lifecycle").get())
+                add("implementation", libs.findLibrary("material").get())
+                add("testImplementation", kotlin("test"))
+            }
         }
+    }
 }
