--- build-logic/convention/src/main/kotlin/com/moya/funch/plugins/AndroidKotlinPlugin.kt
@@ -1,46 +1,0 @@
-package com.moya.funch.plugins
-
-import com.android.build.gradle.BaseExtension
-import com.moya.funch.plugins.utils.libs
-import org.gradle.api.JavaVersion
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.api.plugins.ExtensionAware
-import org.gradle.kotlin.dsl.configure
-import org.gradle.kotlin.dsl.dependencies
-import org.gradle.kotlin.dsl.getByType
-import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
-
-class AndroidKotlinPlugin : Plugin<Project> {
-    override fun apply(target: Project) =
-        with(target) {
-            with(plugins) {
-                apply("kotlin-android")
-            }
-
-            extensions.getByType<BaseExtension>().apply {
-                setCompileSdkVersion(libs.findVersion("compileSdk").get().requiredVersion.toInt())
-
-                defaultConfig {
-                    minSdk = libs.findVersion("minSdk").get().requiredVersion.toInt()
-                    targetSdk = libs.findVersion("targetSdk").get().requiredVersion.toInt()
-                }
-
-                compileOptions {
-                    isCoreLibraryDesugaringEnabled = true
-                    sourceCompatibility = JavaVersion.VERSION_17
-                    targetCompatibility = JavaVersion.VERSION_17
-                }
-
-                (this as ExtensionAware).configure<KotlinJvmOptions> {
-                    jvmTarget = "17"
-                }
-            }
-
-            dependencies {
-                add("coreLibraryDesugaring", libs.findLibrary("desugarLibs").get())
-                add("implementation", libs.findLibrary("kotlin").get())
-                add("implementation", libs.findLibrary("kotlin.coroutines.android").get())
-            }
-        }
-}
